+++ cmd/yb/exec.go
@@ -7,13 +7,10 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/yourbase/yb"
-	"github.com/yourbase/yb/internal/biome"
 	"github.com/yourbase/yb/internal/build"
 	"github.com/yourbase/yb/internal/ybdata"
 	"zombiezen.com/go/log"
 )
-
-const defaultExecEnvironment = "default"
 
 type execCmd struct {
 	execEnvName string
@@ -37,7 +34,7 @@ func newExecCmd() *cobra.Command {
 	envFlagsVar(c.Flags(), &b.env)
 	netrcFlagVar(c.Flags(), &b.netrcFiles)
 	// TODO(light): Use a less confusing name for this flag when it is using targets.
-	c.Flags().StringVar(&b.execEnvName, "environment", defaultExecEnvironment, "Environment to run as")
+	c.Flags().StringVar(&b.execEnvName, "environment", yb.DefaultExecEnvironment, "Environment to run as")
 	return c
 }
 
@@ -65,17 +62,25 @@ func (b *execCmd) run(ctx context.Context) error {
 	if err != nil {
 		return err
 	}
-	bio, err := newBiome(ctx, newBiomeOptions{
+	execTarget := pkg.ExecEnvironments[b.execEnvName]
+	if execTarget == nil {
+		return fmt.Errorf("exec %s: no such environment", b.execEnvName)
+	}
+	biomeOpts := newBiomeOptions{
 		packageDir:      pkg.Path,
-		target:          b.execEnvName,
+		target:          execTarget.Name,
 		dataDirs:        dataDirs,
 		downloader:      downloader,
 		baseEnv:         baseEnv,
 		netrcFiles:      b.netrcFiles,
 		dockerClient:    dockerClient,
-		targetContainer: pkg.Manifest.Exec.Container.ToNarwhal(),
+		targetContainer: execTarget.Container,
 		dockerNetworkID: dockerNetworkID,
-	})
+	}
+	if execTarget.HostOnly {
+		biomeOpts = biomeOpts.disableDocker()
+	}
+	bio, err := newBiome(ctx, biomeOpts)
 	if err != nil {
 		return err
 	}
@@ -92,35 +97,7 @@ func (b *execCmd) run(ctx context.Context) error {
 		Stdout:          os.Stdout,
 		Stderr:          os.Stderr,
 	}
-	defaultEnv, err := biome.MapVars(pkg.Manifest.Exec.Environment[defaultExecEnvironment])
-	if err != nil {
-		return err
-	}
-	deps := &build.PhaseDeps{
-		TargetName:          b.execEnvName,
-		Resources:           narwhalContainerMap(pkg.Manifest.Exec.Dependencies.Containers),
-		EnvironmentTemplate: defaultEnv,
-	}
-	for _, dep := range pkg.Manifest.Dependencies.Runtime {
-		spec, err := yb.ParseBuildpackSpec(dep)
-		if err != nil {
-			return fmt.Errorf("runtime dependencies: %w", err)
-		}
-		deps.Buildpacks = append(deps.Buildpacks, spec)
-	}
-	if b.execEnvName != defaultExecEnvironment {
-		overrideEnv, err := biome.MapVars(pkg.Manifest.Exec.Environment[b.execEnvName])
-		if err != nil {
-			return err
-		}
-		if len(overrideEnv) > 0 && deps.EnvironmentTemplate == nil {
-			deps.EnvironmentTemplate = make(map[string]string)
-		}
-		for k, v := range overrideEnv {
-			deps.EnvironmentTemplate[k] = v
-		}
-	}
-	execBiome, err := build.Setup(ctx, sys, deps)
+	execBiome, err := build.Setup(ctx, sys, execTarget)
 	if err != nil {
 		return err
 	}
@@ -130,9 +107,5 @@ func (b *execCmd) run(ctx context.Context) error {
 			log.Errorf(ctx, "Clean up environment %s: %v", b.execEnvName, err)
 		}
 	}()
-
-	return build.Execute(ctx, sys, &build.Phase{
-		TargetName: b.execEnvName,
-		Commands:   pkg.Manifest.Exec.Commands,
-	})
+	return build.Execute(ctx, sys, execTarget)
 }